* commit '3920e40cd102df8a08893011d85359afb40ed542':
  test fix and cleanup

+++ src/main/java/gov/usgs/cida/pubs/busservice/PersonContributorBusService.java
@@ -83,9 +83,13 @@ public class PersonContributorBusService extends BusService<PersonContributor<?>
 	}
 
 	private void updateAffiliations(Integer contributorId, PersonContributor<?> object) {
-		PersonContributor.getDao().removeAffiliations(contributorId);
-		for (Affiliation<?> affiliation : object.getAffiliations()) {
-			PersonContributor.getDao().addAffiliation(contributorId, affiliation.getId());
+		if (null != object) {
+			PersonContributor.getDao().removeAffiliations(contributorId);
+			if (null != object.getAffiliations()) {
+				for (Affiliation<?> affiliation : object.getAffiliations()) {
+					PersonContributor.getDao().addAffiliation(contributorId, affiliation.getId());
+				}
+			}
 		}
 	}
 }
